trying out better color pallete

+++ include/FractalBase.hpp
@@ -15,7 +15,7 @@ public:
     virtual void compute() = 0;
 
     /* Returns iteration count of a single point, or -1 if inside set */
-    virtual int computePoint(double x, double y) const = 0;
+    virtual double computePoint(double x, double y) const = 0;
 
     void backupAndReplacePointers(sf::Image *newImage, Viewport *newVp);
     void restoreBackedUpPointers();
